Background now becomes transparent once search button is clicked. Foreground search bar UI will be added on top of this.
Created a useState hook to hold the state of the search button click and also created a useState to hold the state of the style for the App.jsx. The setStyle function is now passed as a prop in the SearchBar component to be used to set the style for App.jsx when the button click state changes.